Create Computing Distance between marker edges
It computes the distances between the edges of the marker by finding the shortest distance from a point on one edge to the other edge and then integrates them to find the complete distance between the markers. Then finds the nearest neighbor according to the minimum distance obtained from the markers.